Fix for directory duplication
This change makes it so if you duplicate a folder, the asset references between resources in that folder will point to the new duplicated files.

For instance, if you had dir-a/foo.tscn and dir-a/foo.png where your scene references the png. Previously duplicate would copy both of them to dir-b but dir-b/foo.tscn would still be referencing dir-a/foo.png!